+++ src/mdbx_load.c
@@ -673,7 +673,7 @@ int main(int argc, char *argv[]) {
     goto env_close;
   }
 
-  kbuf.iov_len = mdbx_env_get_maxvalsize_ex(env, 0) + 1;
+  kbuf.iov_len = mdbx_env_get_maxvalsize_ex(env, 0) + (size_t)1;
   if (kbuf.iov_len >= INTPTR_MAX / 2) {
     if (!quiet)
       fprintf(stderr, "mdbx_env_get_maxkeysize() failed, returns %zu\n",
